background

+++ src/utils/vlcManager.js
@@ -12,7 +12,7 @@ const { ipcRenderer } = require('electron');
 const { exec, spawn } = require('child_process');
 const { deleteFileToTrash } = require('./fileExplorer');
 const fs = require('fs').promises, path = require('path');
-const { runPowerShellFile, isExeRunningOnWindows } = require('./childProcess');
+const { runPowerShellFile, isExeRunningOnWindows, timeOutPromise } = require('./childProcess');
 
 const vlc = new VLC.Client({ ip: 'localhost', port: 5029, username: '', password: 'pass' });
 
@@ -22,7 +22,7 @@ const playlistDir = path.join(homedir, 'Documents', 'appsAndMore', 'mySongs', 'p
 
 const checkVLCAndProceed = async (action) => {
 
-    if (!(await isVlcClientRunning())) {
+    if (!(await getVlcClientMode())) {
 
         const notificationInfo = {
 
@@ -40,23 +40,25 @@ const checkVLCAndProceed = async (action) => {
 };
 
 
-const killAllVlcPortableInstances = async () => {
+const killAllVlcPortableInstancesIfAny = async () => {
 
     const execAsync = util.promisify(exec);
+    const isKillNeeded = await isExeRunningOnWindows(VLC_NAME);
 
-    try {
+    if (!isKillNeeded) {
+        return `No ${VLC_NAME} instances found to close it.`;
+    };
 
-        await execAsync(`taskkill /IM ${VLC_NAME} /F`);
+    await execAsync(`taskkill /IM ${VLC_NAME} /F`);
 
-        if (await isExeRunningOnWindows(VLC_NAME)) {
-            throw new Error(`Error closing existing VLC instances : VLC is still running .`);
-        }
-
-        return `All existing ${VLC_NAME} instances closed successfully .`;
-    } catch (err) {
-        console.error(err.message);
-        throw err;
+    const isStillRunningAfterKill = await isExeRunningOnWindows(VLC_NAME);
+    if (isStillRunningAfterKill) {
+        console.error(`Error closing existing VLC instances : VLC is still running .`);
+        throw new Error('killAllVlcPortableInstancesIfAny Failed !');
     }
+
+    return `All existing ${VLC_NAME} instances closed successfully .`;
+
 };
 
 
@@ -68,74 +70,73 @@ const startAndExposeVlcPortable = async () => {
         '--http-port=5029', '--http-password=pass',
         '--qt-start-minimized', '--qt-notification=0'];
 
-    console.log('Starting VlcPortable after all checks and closing existing instances ..');
+    console.log('After all checks and closing Any existing instances ..');
+    console.log('Starting and expose VlcPortable now ..');
 
     const vlcProcess = spawn(vlcExePath, vlcParams,
         { detached: true, stdio: 'ignore' });
 
     vlcProcess.unref();
 
-    await new Promise(resolve => setTimeout(resolve, 2500));
+    await new Promise(resolve => setTimeout(resolve, 1000));
 
-    const isRunning = await isExeRunningOnWindows(VLC_NAME);
+    const isVlcInstanceRunning = await isExeRunningOnWindows(VLC_NAME);
 
-    if (!isRunning) {
+    if (!isVlcInstanceRunning) {
         console.error('Failed to start VlcPortable from startAndExposeVlcPortable ..');
-        throw new Error('VLC failed to start ..');
+        throw new Error('VlcPortable failed to start ..');
     }
 
     console.log('VlcPortable started successfully !');
     return;
-
-
 };
 
 
 
 const initAndRunPlaylistFlow = async () => {
 
+    const soundsDir = path.join(process.env.ASSETS_DIR, 'sound');
+    const createPlaylistScriptPash = path.join(playlistDir, '!create_random_playlist.ps1');
+
+    const randomIndex = Math.floor(Math.random() * 2) + 1;
+    const notificationSound = document.getElementById('notificationSound');
+    notificationSound.src = path.join(soundsDir, `princessPeachRescued${randomIndex}.mp3`);;
+    notificationSound.playbackRate = 1.4;
+    // const soundSrcOnDone = `./../../assets/sound/princessPeachRescued${randomIndex}.mp3`;
+
     try {
 
-        if (await isVlcClientRunning()) {
+        const vlcClientState = await Promise.race([timeOutPromise(), getVlcClientMode()]);
+        const isVlcUnderControl = vlcClientState !== 'timeout' && vlcClientState !== 'unknown';
+
+
+        if (isVlcUnderControl) {
 
             const result = await ipcRenderer.invoke('openDialog', {
                 type: 'question',
-                title: 'VLC is already running ..',
-                message: `What would you like to do ? `,
-                detail: 'Restart it or Keep with the current play ?',
                 buttons: ['Keep', 'Restart',],
+                message: `What would you like to do ? `,
+                title: 'VLC Client is already running ..',
+                detail: 'restart it or keep with the current play ?',
             })
 
             if (result.response !== 1) {
                 console.log('Skip Restarting VLC ...');
                 return;
             }
-        }
+        };
 
-        const randomIndex = Math.floor(Math.random() * 3) + 1;
-        const notificationSound = document.getElementById('notificationSound');
-        const soundSrcOnDone = `./../../assets/sound/princessPeachRescued${randomIndex}.mp3`;
-        const updatePlaylistPs1Path = path.join(playlistDir, '!create_playlist.ps1');
-        const isOtherVlcPortableExeOn = await isExeRunningOnWindows(VLC_NAME);
-        console.log(`Init : isAnyVlcPortableExeOn ${isOtherVlcPortableExeOn}`);
 
-        if (isOtherVlcPortableExeOn) {
-            await killAllVlcPortableInstances();
-        }
-
+        await killAllVlcPortableInstancesIfAny();
         await startAndExposeVlcPortable();
-        await runPowerShellFile(updatePlaylistPs1Path);
-        await addWplFileToPlaylist();
+        await runPowerShellFile(createPlaylistScriptPash);
+        await addWplFileToPlaylist('foreign.wpl');
 
-        notificationSound.src = soundSrcOnDone;
-        notificationSound.play();
-        notificationSound.playbackRate = 1.3;
-
+        await notificationSound.play();
         notificationSound.addEventListener('ended', async () => {
-            await vlc.togglePlay()
+            await vlc.togglePlay();
+            if (await vlc.isPaused()) { await vlc.play() };
         });
-
-
 
     } catch (error) {
 
@@ -143,38 +144,34 @@ const initAndRunPlaylistFlow = async () => {
 
         const notificationInfo = {
             title: 'Error starting VLC !',
-            message: 'Please check the logs for more details ..',
+            message:
+                `Error starting VLC : ${error.message} ..` || 'Unknown error occurred ..',
             icon: 'error',
             sound: true,
-            timeout: 5,
+            timeout: 7,
         }
 
         ipcRenderer.invoke('notificationWIthNode', notificationInfo);
         return;
-
     }
 };
 
 
-const isVlcClientRunning = () => {
+const getVlcClientMode = async () => {
 
-    return new Promise(async (resolve) => {
 
-        try {
+    try {
 
-            const playing = await vlc.isPlaying();
-            const paused = await vlc.isPaused();
+        const paused = await vlc.isPaused();
+        const playing = await vlc.isPlaying();
 
-            if (playing) {
-                resolve('playing');
-            }
-            if (paused) {
-                resolve('paused');
-            }
-        } catch (error) {
-            resolve(false);
-        }
-    });
+        if (paused) return 'paused';
+        if (playing) return 'playing';
+        return 'unknown';
+    } catch (error) {
+        // console.error(`Error checking VLC status : ${error}`);
+        return 'unknown';
+    }
 };
 
 const deleteCurrentSongAndPlayNext = async () => {
@@ -196,8 +193,10 @@ const deleteCurrentSongAndPlayNext = async () => {
             const soundSrcOnDone = `./../../assets/sound/macEmptyTrash.mp3`;
             notificationSound.src = soundSrcOnDone;
             notificationSound.play();
-            await new Promise(resolve => setTimeout(resolve, 1500));
-            await pauseOrResume();
+            await new Promise(resolve => setTimeout(resolve, 1000));
+            await vlc.togglePlay();
+            await new Promise(resolve => setTimeout(resolve, 500));
+            await vlc.play();
             return;
         } catch (error) {
             console.error(`Error deleting current song and playing next : ${error}`);
@@ -209,80 +208,66 @@ const deleteCurrentSongAndPlayNext = async () => {
 
 const loveThisSong = async () => {
 
+    const currentVolume = await vlc.getVolume();
     const baseFolder = path.join(homedir, 'Documents', 'appsAndMore');
     const loveThisSongsFolder = path.join(baseFolder, 'mySongs', 'rest', 'loveThisSongs');
 
-    return checkVLCAndProceed(async () => {
-
-        try {
-
-            if (!require('fs').existsSync()) {
-                await fs.mkdir(loveThisSongsFolder, { recursive: true })
-            };
-
-            const playlist = await vlc.getPlaylist();
-            const currentEntry = playlist.find(entry => entry.isCurrent);
-            const fullPath = currentEntry.uri.replace('file:///', '');
-            console.log(`fullPath Song : ${fullPath}`);
-            const fileName = path.basename(fullPath);
-            const newFilePath = path.join(loveThisSongsFolder, fileName);
-            await fs.copyFile(fullPath, newFilePath);
-            const notificationSound = document.getElementById('notificationSound');
-            const soundSrcOnDone = `./../../assets/sound/successSound.mp3`;
-            notificationSound.src = soundSrcOnDone;
-            await pauseOrResume();
-            notificationSound.play();
-            await new Promise(resolve => setTimeout(resolve, 1500));
-            await pauseOrResume();
-            console.log(`Song copied to loveThisSongs folder : ${newFilePath}`);
-            return;
 
 
-        } catch (error) {
-            console.error(`Error copied Song : ${error}`);
-            return;
-        }
-    });
+    if (!require('fs').existsSync()) {
+        await fs.mkdir(loveThisSongsFolder, { recursive: true })
+    };
+
+    const playlist = await vlc.getPlaylist();
+    const currentEntry = playlist.find(entry => entry.isCurrent);
+    const fullPath = currentEntry.uri.replace('file:///', '');
+    console.log(`fullPath Song : ${fullPath}`);
+    const fileName = path.basename(fullPath);
+    const newFilePath = path.join(loveThisSongsFolder, fileName);
+    await fs.copyFile(fullPath, newFilePath);
+    const notificationSound = document.getElementById('notificationSound');
+    const soundSrcOnDone = `./../../assets/sound/successSound.mp3`;
+    notificationSound.src = soundSrcOnDone;
+    await vlc.setVolume((currentVolume / 2));
+    notificationSound.play();
+    await new Promise(resolve => setTimeout(resolve, 1300));
+    await vlc.setVolume(currentVolume);
+    console.log(`Song copied to loveThisSongs folder : ${newFilePath}`);
+    return;
 };
 
 
 const addWplFileToPlaylist = async (wplFileName = 'foreign.wpl') => {
 
     const playlistFilePath = path.join(playlistDir, wplFileName);
-    console.log('Attempting to add playlist to VLC ...');
     await vlc.addToPlaylist(playlistFilePath);
-    console.log(`Playlist added : ${playlistFilePath}`);
-    await new Promise(resolve => setTimeout(resolve, 1000));
-    // return;
+    console.log(`PlaylistFile ${wplFileName} added successfully !`);
+    // await new Promise(resolve => setTimeout(resolve, 500));
+    return;
 };
 
 const pauseOrResume = async () => {
 
-    return checkVLCAndProceed(async () => {
+    // return checkVLCAndProceed(async () => {
 
-        try {
-
-            // setTimeout(() => {
-            //     console.log('Timeout of 4 seconds - reject it ..');
-            //     return 'paused';
-            // }, 3000);
-            const isPlaying = await vlc.isPlaying();
-            const isPaused = await vlc.isPaused();
-            if (isPlaying) {
-                await vlc.pause();
-                console.log('Playback paused .');
-                return 'paused';
-            } if (isPaused) {
-                await vlc.play();
-                console.log('Playback resumed .');
-                return 'play';
-            }
-            return;
-        } catch (error) {
-            console.error(`Error pausing or resuming playback : ${error}`);
-            return;
+    try {
+        const isPlaying = await vlc.isPlaying();
+        const isPaused = await vlc.isPaused();
+        if (isPlaying) {
+            await vlc.pause();
+            console.log('Playback paused .');
+            return 'paused';
+        } if (isPaused) {
+            await vlc.play();
+            console.log('Playback resumed .');
+            return 'play';
         }
-    });
+        return;
+    } catch (error) {
+        console.error(`Error pausing or resuming playback : ${error}`);
+        return;
+    }
+    // });
 };
 
 const playNext = () => {
@@ -314,7 +299,7 @@ const playPrevious = () => {
 module.exports = {
     playNext, playPrevious, loveThisSong,
     initAndRunPlaylistFlow, pauseOrResume,
-    deleteCurrentSongAndPlayNext, isVlcClientRunning
+    deleteCurrentSongAndPlayNext, getVlcClientMode
 };
